Fixed signed/unsigned comparison warnings in empty path gm/sample
git-svn-id: http://skia.googlecode.com/svn/trunk@2439 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ gm/emptypath.cpp
@@ -83,8 +83,8 @@ protected:
         canvas->save();
         canvas->translate(10 * SK_Scalar1, 0);
         canvas->save();
-        for (int style = 0; style < SK_ARRAY_COUNT(gStyles); ++style) {
-            for (int fill = 0; fill < SK_ARRAY_COUNT(gFills); ++fill) {
+        for (size_t style = 0; style < SK_ARRAY_COUNT(gStyles); ++style) {
+            for (size_t fill = 0; fill < SK_ARRAY_COUNT(gFills); ++fill) {
                 if (0 == i % 4) {
                     canvas->restore();
                     canvas->translate(0, rect.height() + 40 * SK_Scalar1);
